Ppr/PPM/CPMMonitoring: Fix wrong variable type for event numbers and bug related to mismatch-event-counter
Ppr/PPM/CPMMonitoring: Fix wrong variable type for event numbers.

The event number variable used in the Ppr-,PPM, and CPM-Monitoring was inialized as integer, whereas it should be long long.
This lead to negative event numbers (2-complement) or none at all.

The bug was fixed by changing the type.

+++ Trigger/TrigT1/TrigT1CaloMonitoring/src/CpmSimMonitorAlgorithm.cxx
@@ -253,7 +253,7 @@ StatusCode CpmSimMonitorAlgorithm::fillHistograms( const EventContext& ctx ) con
 	cpmError=err;
 	fill(m_packageName,cpmErrorLoc,cpmError);
       }
-      const int eventNumber = ctx.eventID().event_number();
+      const long long eventNumber = ctx.eventID().event_number();
       if ((errorsCPM[loc + cpmBins] >> err) & 0x1) {
 	cpmErrorLoc_SimNeData=loc;
 	cpmError_SimNeData=err;
